@@ -1,6 +1,6 @@
 {
   "name": "web",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve --live-reload true --proxy-config proxy.json",
@@ -11,47 +11,53 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.0.0",
-    "@angular/cdk": "^8.0.0",
-    "@angular/common": "^8.0.0",
-    "@angular/compiler": "^8.0.0",
-    "@angular/core": "^8.0.0",
-    "@angular/forms": "^8.0.0",
+    "@angular/animations": "^8.2.13",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "^8.2.13",
+    "@angular/compiler": "^8.2.13",
+    "@angular/core": "^8.2.13",
+    "@angular/forms": "^8.2.13",
     "@angular/http": "^7.2.15",
-    "@angular/material": "^8.0.0",
-    "@angular/material-moment-adapter": "^8.0.0",
-    "@angular/platform-browser": "^8.0.0",
-    "@angular/platform-browser-dynamic": "^8.0.0",
-    "@angular/pwa": "^0.800.1",
-    "@angular/router": "^8.0.0",
-    "@angular/service-worker": "^8.0.0",
+    "@angular/material": "^8.2.3",
+    "@angular/material-moment-adapter": "^8.2.3",
+    "@angular/platform-browser": "^8.2.13",
+    "@angular/platform-browser-dynamic": "^8.2.13",
+    "@angular/pwa": "^0.800.6",
+    "@angular/router": "^8.2.13",
+    "@angular/service-worker": "^8.2.13",
     "@manekinekko/angular-web-bluetooth": "^7.0.0",
-    "core-js": "^3.1.3",
+    "core-js": "^3.3.6",
+    "g": "^2.0.1",
     "material-design-icons": "^3.0.1",
     "moment": "^2.24.0",
-    "rxjs": "^6.5.2",
-    "typescript": "^3.4.5",
+    "ngx-toastr": "^10.2.0",
+    "rxjs": "^6.5.3",
+    "typescript": "^3.5.3",
     "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.800.1",
-    "@angular/cli": "^8.0.1",
-    "@angular/compiler-cli": "^8.0.0",
-    "@angular/language-service": "^8.0.0",
-    "@types/jasmine": "^3.3.13",
-    "@types/jasminewd2": "^2.0.6",
-    "@types/node": "^12.0.4",
+    "@angular-devkit/build-angular": "^0.800.6",
+    "@angular/cli": "^8.3.17",
+    "@angular/compiler-cli": "^8.2.13",
+    "@angular/language-service": "^8.2.13",
+    "@types/jasmine": "^3.4.5",
+    "@types/jasminewd2": "^2.0.8",
+    "@types/node": "^12.12.5",
     "@types/web-bluetooth": "0.0.4",
     "codelyzer": "~5.0.1",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^4.1.0",
+    "karma": "^4.4.1",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^2.0.5",
+    "karma-coverage-istanbul-reporter": "^2.1.0",
     "karma-jasmine": "~2.0.1",
     "karma-jasmine-html-reporter": "^1.4.2",
     "protractor": "^5.4.2",
     "ts-node": "~8.2.0",
     "tslint": "~5.17.0"
-  }
+  },
+  "browserslist": [
+    "last 2 version",
+    "> 1%"
+  ]
 }
