@@ -1,10 +1,11 @@
 package sendmail
 
 import (
-	"errors"
+	"bytes"
 	"fmt"
+	"html/template"
 	"io/ioutil"
-	"sto/server/config"
+	"odo24/server/config"
 
 	email "github.com/mil-ast/sendmail"
 )
@@ -13,6 +14,7 @@ import (
 const (
 	TypeConfirmEmail uint8 = iota
 	TypeRepairConfirmCode
+	TypeRegisterFromOAUTH
 )
 
 var templates map[uint8]string
@@ -39,11 +41,11 @@ func init() {
 	templates[TypeRepairConfirmCode] = string(body)
 }
 
-// SendEmail
-func SendEmail(to string, tpl uint8, code uint32) error {
-	body, ok := templates[tpl]
+// SendEmail отправка
+func SendEmail(to string, tplID uint8, params map[string]interface{}) error {
+	templateBody, ok := templates[tplID]
 	if !ok {
-		return errors.New("Template not founc")
+		return fmt.Errorf("Template %d not found", tplID)
 	}
 
 	options := config.GetInstance()
@@ -60,8 +62,16 @@ func SendEmail(to string, tpl uint8, code uint32) error {
 		return err
 	}
 
-	message := fmt.Sprintf(body, options.App.SmtpFrom, to, code)
-	err = client.Send(options.App.SmtpFrom, to, message)
+	templateBody = fmt.Sprintf(templateBody, options.App.SmtpFrom, to)
+
+	buffer := new(bytes.Buffer)
+	t := template.Must(template.New("letter").Parse(templateBody))
+	err = t.Execute(buffer, params)
+	if err != nil {
+		return err
+	}
+
+	err = client.Send(options.App.SmtpFrom, to, buffer.String())
 	if err != nil {
 		return err
 	}
