@@ -1,201 +1,86 @@
 package sessions
 
 import (
-	"crypto/rand"
-	"encoding/json"
+	"encoding/hex"
 	"errors"
-	"fmt"
-	"log"
 	"net/http"
-	"sto/server/config"
-	"sto/server/models"
+	"odo24/server/api/models"
 	"time"
 
-	"bitbucket.org/mil-ast/memcached"
-	"github.com/mil-ast/db"
+	"github.com/gin-gonic/gin"
 )
 
 const (
-	SessionSize        int           = 32
-	SessionID          string        = "odo.s"
-	SessionTimeLife    time.Duration = time.Hour * 24 * 91
-	SessionMemcTimeout time.Duration = time.Hour
-	charts             string        = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
+	sessionID      string        = "sess"
+	SessionTimeout time.Duration = time.Hour * time.Duration(24*31)
 )
 
-var memc *memcached.Memcached
-
-func init() {
-	cfg := config.GetInstance()
-	var err error
-	memc, err = memcached.NewConn(cfg.App.MemcachedAddr)
-	if err != nil {
-		panic(err)
-	}
-}
-
-func GetSession(w http.ResponseWriter, r *http.Request) (models.Profile, error) {
-	var profile models.Profile
-
-	cookie, err := r.Cookie(SessionID)
-	if err != nil {
-		log.Println(err)
-		return profile, err
+func NewSession(c *gin.Context, userID uint64) error {
+	if userID == 0 {
+		return errors.New("profile is empty")
 	}
 
-	data, err := memc.Get(cookie.Value)
-	if err != nil {
-		// если ошибки с Memcached, смотрим сессию в БД
-		if err != nil {
-			userAgent := r.Header.Get("user-agent")
-			profile, err = getSessionFromDB(cookie.Value, userAgent)
-			if err != nil {
-				return profile, err
-			}
-
-			return profile, nil
-		}
-
-		return profile, err
+	sess := models.SessionValue{
+		UserID:     userID,
+		Expiration: uint64(time.Now().Add(SessionTimeout).UTC().Unix()),
 	}
 
-	err = json.Unmarshal(data, &profile)
-	if err != nil {
-		return profile, err
-	}
-
-	err = UpdateSession(w, r, profile)
-	return profile, err
-}
-
-func getProfileByUUID(uuid string) (models.Profile, error) {
-	var profile models.Profile
-	return profile, nil
-}
-
-func UpdateSession(w http.ResponseWriter, r *http.Request, profile models.Profile) error {
-	cookie, err := r.Cookie(SessionID)
+	sessionValue, err := encrypt(sess.Bytes())
 	if err != nil {
 		return err
+	}
+
+	c.SetCookie(sessionID, hex.EncodeToString(sessionValue), int(SessionTimeout.Seconds()), "/", "", false, true)
+
+	return nil
+}
+
+func GetSession(c *gin.Context) (*models.SessionValue, error) {
+	cookie, err := c.Request.Cookie(sessionID)
+	if err != nil {
+		return nil, err
 	}
 
 	if cookie.Value == "" {
-		return errors.New("empty session id")
+		return nil, errors.New("expired")
 	}
 
-	data, err := json.Marshal(profile)
+	decodedStr, err := hex.DecodeString(cookie.Value)
 	if err != nil {
-		return err
+		DeleteSession(c)
+		return nil, err
 	}
 
-	expires := time.Now().Add(SessionTimeLife)
-	cookie.Expires = expires
+	value, err := decrypt(decodedStr)
+	if err != nil {
+		DeleteSession(c)
+		return nil, errors.New("expired")
+	}
 
-	return memc.Set(cookie.Value, SessionMemcTimeout, data)
+	sess := new(models.SessionValue)
+	err = sess.Parse(value)
+	if err != nil {
+		DeleteSession(c)
+		return nil, err
+	}
+
+	now := time.Now().UTC().Unix()
+	if sess.Expiration < uint64(now) {
+		return nil, errors.New("expired")
+	}
+
+	return sess, nil
 }
 
-func NewSession(w http.ResponseWriter, r *http.Request, profile models.Profile) error {
-	cookie, err := r.Cookie(SessionID)
-	if err == nil {
-		memc.Delete(cookie.Value)
+// DeleteSession удаление сессии
+func DeleteSession(c *gin.Context) {
+	cookie := &http.Cookie{
+		Name:     sessionID,
+		Value:    "",
+		Path:     "/",
+		Expires:  time.Unix(0, 0),
+		HttpOnly: true,
 	}
 
-	sesID := createSessionID()
-	expires := time.Now().Add(SessionTimeLife)
-	cookie = &http.Cookie{Name: SessionID, Value: sesID, Path: "/", HttpOnly: true, Expires: expires}
-	http.SetCookie(w, cookie)
-
-	data, err := json.Marshal(profile)
-	if err != nil {
-		return err
-	}
-
-	err = memc.Add(cookie.Value, SessionMemcTimeout, data)
-	if err != nil {
-		return err
-	}
-
-	userAgent := r.Header.Get("user-agent")
-
-	err = saveSessionToDB(cookie.Value, userAgent, profile)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func DelSession(w http.ResponseWriter, r *http.Request) error {
-	cookie, err := r.Cookie(SessionID)
-	if err != nil {
-		return err
-	}
-
-	key := cookie.Value
-
-	log.Println(777, key)
-	memc.Delete(key)
-	deleteSessionFromoDB(key)
-
-	expires := time.Now()
-	cookie = &http.Cookie{Name: SessionID, Value: "", HttpOnly: true, Expires: expires}
-	http.SetCookie(w, cookie)
-
-	return nil
-}
-
-func createSessionID() string {
-	b := make([]byte, 16)
-	_, err := rand.Read(b)
-	if err != nil {
-		return ""
-	}
-
-	return fmt.Sprintf("%X-%X-%X-%X-%X", b[0:4], b[4:6], b[6:8], b[8:10], b[10:])
-
-	/*output := make([]byte, SessionSize)
-	randomness := make([]byte, SessionSize)
-	_, err := rand.Read(randomness)
-	if err != nil {
-		return ""
-	}
-	l := len(charts)
-	for pos := range output {
-		random := uint8(randomness[pos])
-		randomPos := random % uint8(l)
-		output[pos] = charts[randomPos]
-	}
-	return string(output)*/
-}
-
-func getSessionFromDB(uuid, userAgent string) (models.Profile, error) {
-	var profile models.Profile
-
-	conn, err := db.GetConnection()
-	if err != nil {
-		return profile, err
-	}
-
-	row := conn.QueryRow("select user_id,login,is_no_confirmed from profiles.getprofilebyuuid($1,$2)", uuid, userAgent)
-	err = row.Scan(&profile.User_id, &profile.Login, &profile.IsNoConfirmed)
-	return profile, err
-}
-
-func saveSessionToDB(uuid, userAgent string, profile models.Profile) error {
-	conn, err := db.GetConnection()
-	if err != nil {
-		return err
-	}
-
-	query := fmt.Sprintf("select profiles.profileaddsession($1::bigint,$2::uuid,$3::varchar,now()::timestamp + interval '%0.f hours')", SessionTimeLife.Hours())
-	_, err = conn.Exec(query, profile.User_id, uuid, userAgent)
-	return err
-}
-
-func deleteSessionFromoDB(uuid string) error {
-	conn, err := db.GetConnection()
-	if err != nil {
-		return err
-	}
-	_, err = conn.Exec("select profiles.profiledelsession($1::uuid)", uuid)
-	return err
+	http.SetCookie(c.Writer, cookie)
 }
