@@ -1,15 +1,11 @@
 export class Profile {
   user_id: number;
   login: string;
-  phone: number;
-  time_zone: number;
-  is_no_confirmed: boolean;
+  confirmed: boolean;
 
   constructor(data: any) {
     this.user_id = data.user_id | 0;
     this.login = data.login || null;
-    this.phone = data.phone | 0;
-    this.time_zone = data.time_zone | 10;
-    this.is_no_confirmed = data.is_no_confirmed || false;
+    this.confirmed = !!data.confirmed;
   }
 }
