@@ -4,20 +4,18 @@ import { Observable, BehaviorSubject } from 'rxjs';
 import { map } from 'rxjs/operators';
 
 export interface GroupStruct {
-    group_id?: number;
-    name: string;
-    order?: number;
-    global?: boolean;
-    cnt?: number;
+  group_id: number;
+  group_name: string;
+  sort: number;
 }
-export interface GroupsStats {
-    group_id?: number;
-    cnt?: number;
+export interface GroupStructModify {
+  group_name: string;
 }
 
 @Injectable()
 export class GroupService {
-  private url = '/api/groups';
+  private readonly url = '/api/groups';
+  private readonly urlGroup = '/api/group';
 
   private selectedGroup: BehaviorSubject<GroupStruct> = new BehaviorSubject(null);
   selected: Observable<GroupStruct> = this.selectedGroup.asObservable();
@@ -26,58 +24,47 @@ export class GroupService {
     private http: HttpClient
   ) { }
 
-  get(avtoId: number = 0): Observable<GroupStruct[]> {
-    return this.http.get<GroupStruct[]>(this.url, {
-      params: {
-        avto_id: `${avtoId}`
+  get(): Observable<GroupStruct[]> {
+    return this.http.get<GroupStruct[]>(this.url).pipe(
+      map((list: GroupStruct[]) => {
+        const sorted = (list || []).sort((a: GroupStruct, b: GroupStruct) => a.sort - b.sort);
+        const selectedGroup = this.selectedGroup.getValue();
+        if (list.length > 0 && selectedGroup === null) {
+          this.setSelected(list[0]);
+        }
+        return sorted;
       }
-    }).pipe(map((list: GroupStruct[]) => {
-      list = list || [];
-      return list.sort((a: GroupStruct, b: GroupStruct) => {
-        if (a.global && !b.global) {
-          return -1;
-        } else if (!a.global && b.global) {
-          return 1;
-        }
-
-        if (a.order > b.order) {
-          return 1;
-        } else if (a.order < b.order) {
-          return -1;
-        }
-        return 0;
-      });
-    }));
+    ));
   }
 
-  getStats(avtoId: number): Observable<GroupsStats[]> {
-    return this.http.get<GroupsStats[]>(`${this.url}/stats`, {
-      params: {
-        avto_id: avtoId.toString()
-      }
-    });
+  saveNewSort(groups: GroupStruct[]): Observable<void> {
+    const body = (groups || []).map((group: GroupStruct) => group.group_id);
+    return this.http.put<void>(`${this.url}/sort`, body);
   }
 
-  create(data: GroupStruct) {
+  create(data: GroupStructModify) {
     return this.http.post(this.url, data);
   }
 
-  update(data: GroupStruct) {
-    return this.http.put(this.url, data);
+  update(groupID: number, data: GroupStructModify) {
+    return this.http.put(`${this.urlGroup}/${groupID}`, data);
   }
 
-  delete(id: number) {
-    return this.http.delete(this.url.concat(`?group_id=${id}`));
+  delete(groupID: number) {
+    return this.http.delete(`${this.urlGroup}/${groupID}`);
   }
 
-  setSelected(avto: GroupStruct) {
-    this.selectedGroup.next(avto);
+  setSelected(group: GroupStruct) {
+    this.selectedGroup.next(group);
+  }
+  getSelected(): GroupStruct {
+    return this.selectedGroup.getValue();
   }
   resetSelected(): void {
     this.selectedGroup.next(null);
   }
 
-  isSelected(avto: GroupStruct): boolean {
-    return this.selectedGroup.getValue() === avto;
+  isSelected(group: GroupStruct): boolean {
+    return this.selectedGroup.getValue() === group;
   }
 }
