@@ -1,14 +1,14 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { HomeComponent } from './home/home.component';
 import { AuthGuard } from './auth.guard';
 import { ServicesComponent } from './services/services.component';
+import { ProfileComponent } from './profile/profile.component';
 
 const appRoutes: Routes = [
-  { path: '', loadChildren: './home/home.module#HomeModule' },
+  { path: 'profile', component: ProfileComponent, loadChildren: () => import('./profile/profile.module').then(m => m.ProfileModule) },
   { path: 'service', component: ServicesComponent, canActivate: [AuthGuard] },
-  { path: 'documents', loadChildren: './documents/documents.module#DocumentsModule', canActivate: [AuthGuard] },
-  { path: '**', redirectTo: '' }
+  { path: 'documents', loadChildren: () => import('./documents/documents.module').then(m => m.DocumentsModule), canActivate: [AuthGuard] },
+  { path: '**', redirectTo: '/profile/login' }
 ];
 
 @NgModule({
