@@ -1,17 +1,23 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ServicesComponent } from './services.component';
-import { MatButtonModule } from '@angular/material/button';
-import { MatCardModule } from '@angular/material/card';
-import { MatDatepickerModule } from '@angular/material/datepicker';
-import { MatDialogModule } from '@angular/material/dialog';
-import { MatIconModule } from '@angular/material/icon';
-import { MatInputModule } from '@angular/material/input';
-import { MatMenuModule } from '@angular/material/menu';
-import { MatProgressBarModule } from '@angular/material/progress-bar';
-import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
-import { MatSnackBarModule } from '@angular/material/snack-bar';
-import { AvtoService } from '../_services/avto.service';
+import {
+  MatButtonModule,
+  MatSidenavModule,
+  MatSnackBarModule,
+  MatProgressSpinnerModule,
+  MatProgressBarModule,
+  MatDialogModule,
+  MatInputModule,
+  MatIconModule,
+  MatMenuModule,
+  MatDatepickerModule,
+  MatToolbarModule,
+  MatSelectModule,
+  MatDividerModule,
+  MatListModule,
+} from '@angular/material';
+import { AutoService } from '../_services/avto.service';
 import { SelectedAvtoModule } from '../shared/selected-avto/selected-avto.module';
 import { GroupService } from '../_services/groups.service';
 import { ItemServiceComponent } from './item-service/item-service.component';
@@ -21,6 +27,15 @@ import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { DialogCreateServiceComponent } from './dialogs/dialog-create-service/dialog-create-service.component';
 import { DialogUpdateServiceComponent } from './dialogs/dialog-update-service/dialog-update-service.component';
 import { OrderPipeModule } from '../_pipes/order.pipe.module';
+import { ListAvtoComponent } from '../shared/list-avto/list-avto.component';
+import { SharedUtilsModule } from '../shared_modules/utils/shared-utils.modue';
+import { GroupsConfigComponent } from './groups-config/groups-config.component';
+import { DragDropModule } from '@angular/cdk/drag-drop';
+import { ServiceExpirationComponent } from './service-expiration/service-expiration.component';
+import { LeftDistancePipe } from './pipes/left-distance.pipe';
+import { DistanceLevelPipe } from './pipes/distance-level.pipe';
+import { DialogUpdateGroupComponent } from './dialogs/dialog-update-group/dialog-update-group.component';
+import { DialogCreateGroupComponent } from './dialogs/dialog-create-group/dialog-create-group.component';
 
 @NgModule({
   declarations: [
@@ -30,11 +45,20 @@ import { OrderPipeModule } from '../_pipes/order.pipe.module';
     DialogCreateAvtoComponent,
     DialogCreateServiceComponent,
     DialogUpdateServiceComponent,
+    DialogUpdateGroupComponent,
+    DialogCreateGroupComponent,
+    ListAvtoComponent,
+    GroupsConfigComponent,
+    ServiceExpirationComponent,
+    LeftDistancePipe,
+    DistanceLevelPipe,
   ],
   imports: [
     CommonModule,
     SelectedAvtoModule,
+    MatSidenavModule,
     MatSnackBarModule,
+    SharedUtilsModule,
     MatProgressSpinnerModule,
     MatProgressBarModule,
     MatDialogModule,
@@ -42,14 +66,18 @@ import { OrderPipeModule } from '../_pipes/order.pipe.module';
     MatInputModule,
     MatIconModule,
     MatMenuModule,
-    MatCardModule,
+    MatSelectModule,
     MatDatepickerModule,
+    MatListModule,
+    MatToolbarModule,
+    MatDividerModule,
+    DragDropModule,
     FormsModule,
     ReactiveFormsModule,
     OrderPipeModule,
   ],
   providers: [
-    AvtoService,
+    AutoService,
     GroupService,
     ServiceService,
   ],
@@ -57,6 +85,8 @@ import { OrderPipeModule } from '../_pipes/order.pipe.module';
     DialogCreateAvtoComponent,
     DialogCreateServiceComponent,
     DialogUpdateServiceComponent,
+    DialogUpdateGroupComponent,
+    DialogCreateGroupComponent,
   ]
 })
 export class ServiceModule {}
