@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { AvtoStruct } from '../../_classes/avto';
+import { AutoStruct } from '../../_classes/auto';
 import { MatDialog } from '@angular/material/dialog';
 import { DialogUpdateAvtoComponent } from 'src/app/shared/dialog-update-avto/dialog-update-avto.component';
 
@@ -9,7 +9,7 @@ import { DialogUpdateAvtoComponent } from 'src/app/shared/dialog-update-avto/dia
   styleUrls: ['./selected-avto.component.scss']
 })
 export class SelectedAvtoComponent {
-  @Input() model: AvtoStruct;
+  @Input() model: AutoStruct;
 
   constructor(
     private dialog: MatDialog,
@@ -25,6 +25,6 @@ export class SelectedAvtoComponent {
   }
 
   get iconURL(): string {
-    return `/api/images/small/${this.model.avto_id}.jpg?t=1`;
+    return `/api/images/small/${this.model.auto_id}.jpg?t=1`;
   }
 }
